refactor: 검색 기능 관련 Product -> Search 패키지로 이동
