+++ src/main/java/com/sellect/server/product/controller/ProductController.java
@@ -2,7 +2,6 @@ package com.sellect.server.product.controller;
 
 import com.sellect.server.auth.domain.User;
 import com.sellect.server.common.infrastructure.annotation.AuthSeller;
-import com.sellect.server.common.infrastructure.annotation.AuthUser;
 import com.sellect.server.common.response.ApiResponse;
 import com.sellect.server.product.application.ProductImageService;
 import com.sellect.server.product.application.ProductService;
@@ -11,16 +10,10 @@ import com.sellect.server.product.controller.request.ProductModifyRequest;
 import com.sellect.server.product.controller.request.ProductRegisterRequest;
 import com.sellect.server.product.controller.response.ProductModifyResponse;
 import com.sellect.server.product.controller.response.ProductRegisterResponse;
-import com.sellect.server.product.domain.Product;
-import com.sellect.server.product.domain.ProductSearchCondition;
-import com.sellect.server.product.domain.ProductSortType;
-import jakarta.servlet.http.HttpServletRequest;
-import jakarta.servlet.http.HttpServletResponse;
 import jakarta.validation.Valid;
 import java.util.List;
 import lombok.RequiredArgsConstructor;
 import org.springframework.web.bind.annotation.DeleteMapping;
-import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PatchMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
@@ -38,7 +31,7 @@ public class ProductController {
     private final ProductService productService;
     private final ProductImageService productImageService;
 
-    /*
+    /**
      * 상품 등록 (복수 지원)
      * */
     @PostMapping("/product")
@@ -78,41 +71,7 @@ public class ProductController {
         return ApiResponse.ok();
     }
 
-    /*
-     * 상품 조회 API
-     * condition 을 통해 확인
-     * */
-    @GetMapping("/products/search")
-    // todo : 브랜드, 리뷰, 이미지 엔티티 생성 후 다시 돌아올 것
-    public List<Product> search(
-        @AuthUser User user,
-        @RequestParam String keyword,
-        @RequestParam(required = false) Long categoryId,
-        @RequestParam(required = false) Long brandId,
-        @RequestParam(required = false) Integer minPrice,
-        @RequestParam(required = false) Integer maxPrice,
-        @RequestParam(defaultValue = "0") int page,
-        @RequestParam(defaultValue = "20") int size,
-        @RequestParam(defaultValue = "LATEST") ProductSortType sortType,
-        @RequestParam(defaultValue = "false") boolean isInitialSearch,
-        HttpServletRequest request,
-        HttpServletResponse response
-    ) {
-        ProductSearchCondition condition = ProductSearchCondition
-            .builder()
-            .keyword(keyword)
-            .categoryId(categoryId)
-            .brandId(brandId)
-            .minPrice(minPrice)
-            .maxPrice(maxPrice)
-            .build();
-
-        Long userId = (user != null) ? user.getId() : null;
-        // todo : ProductReadResponse에 잘 맵핑해서 보낼 것
-        return productService.search(userId, condition, page, size, sortType, isInitialSearch, request, response);
-    }
-
-    /*
+    /**
      * 상품 이미지 수정 API
      * */
     @PostMapping("/products/images")
