+++ src/main/java/com/sellect/server/product/repository/ProductRepository.java
@@ -1,18 +1,12 @@
 package com.sellect.server.product.repository;
 
 import com.sellect.server.product.domain.Product;
-import com.sellect.server.product.domain.ProductSearchCondition;
-import com.sellect.server.product.domain.ProductSortType;
 import java.util.List;
 import java.util.Optional;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 
 public interface ProductRepository {
-
-    // 상품 조회 (동적 쿼리)
-    // todo : 브랜드, 리뷰, 이미지 엔티티 생성 후 다시 돌아올 것
-    List<Product> search(ProductSearchCondition condition, int page, int size, ProductSortType sortType);
 
     List<Product> saveAll(List<Product> products);
 
@@ -25,11 +19,5 @@ public interface ProductRepository {
     Product save(Product product);
 
     Page<Product> findContainingName(String keyword, Pageable pageable);
-
-    Page<Product> findByCategoryId(Long categoryId, Pageable pageable);
-
-    Page<Product> findByBrandId(Long brandId, Pageable pageable);
-
-    Page<Product> findByIdIn(List<Long> ids, Pageable pageable);
 
 }
