+++ src/main/java/com/sellect/server/search/controller/response/SearchResponse.java
@@ -1,10 +1,10 @@
-package com.sellect.server.product.controller.response;
+package com.sellect.server.search.controller.response;
 
 import java.math.BigDecimal;
 import lombok.Builder;
 
 @Builder
-public record ProductSearchResponse(
+public record SearchResponse(
     // todo : 브랜드, 리뷰, 이미지 엔티티 생성 후 다시 돌아올 것
     String brandName,
     String productId,
