+++ src/main/java/com/sellect/server/search/domain/SearchCondition.java
@@ -1,4 +1,4 @@
-package com.sellect.server.product.domain;
+package com.sellect.server.search.domain;
 
 import lombok.AccessLevel;
 import lombok.AllArgsConstructor;
@@ -8,7 +8,7 @@ import lombok.Getter;
 @Getter
 @AllArgsConstructor(access = AccessLevel.PRIVATE) // @Builder 사용 시 명확한 객체 생성
 @Builder
-public class ProductSearchCondition {
+public class SearchCondition {
 
     // todo : 브랜드, 리뷰, 이미지 엔티티 생성 후 다시 돌아올 것
     private final String keyword;
